Fold looping and track information from LoopingPcmStream into PcmStream
